requirements.txt
Django==4.0.4
djangorestframework==3.13.1
channels==3.0.4
redis==4.2.2
channels-redis==3.3.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
uvicorn==0.17.0
